+++ lib/core/services/ledgerparty.dart
@@ -1,7 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:tassist/core/models/ledgerparty.dart';
 
-
 class LedgerPartyService {
   final String uid;
 
@@ -9,11 +8,10 @@ class LedgerPartyService {
 
   LedgerPartyService({this.uid, this.voucherId});
 
-   final CollectionReference companyCollection =
+  final CollectionReference companyCollection =
       Firestore.instance.collection('company');
 
- 
-    Stream<List<LedgerParty>> get voucherLedgerData {
+  Stream<List<LedgerParty>> get voucherLedgerData {
     return companyCollection
         .document(this.uid)
         .collection('ledger_entries')
@@ -22,25 +20,23 @@ class LedgerPartyService {
         .map(_voucherLedgerData);
   }
 
-  
-
-
   List<LedgerParty> _voucherLedgerData(QuerySnapshot snapshot) {
     return snapshot.documents.map((doc) {
       return LedgerParty(
-          amount: doc.data['amount'].toInt() ?? 0,
-          isDeemedPositive: doc.data['isdeemedpositive'].toString() ?? '',
-          isPartyLedger: doc.data['ispartyledger'].toString() ?? '',
-          ledgerGuid: doc.data['ledger_guid'].toString() ?? '',
-          ledgerMasterId: doc.data['ledgername'].toString() ?? '',
-          ledgerRefMasterId: doc.data['ledgerrefname'].toString() ?? '',
-          primaryVoucherType: doc.data['primary_voucher_type_name'].toString() ?? '',
-          primaryGroup: doc.data['primarygroup'].toString() ?? '',
-          ledgerName: doc.data['restat_ledger_name'].toString() ?? '',
-          ledgerRefName: doc.data['restat_ledger_ref_name'].toString() ?? '',
-          partyName: doc.data['restat_party_ledger_name'].toString() ?? '',
-          date: doc.data['voucher_date'].toDate() ?? null,
-          voucherMasterID: doc.data['voucher_master_id'].toString() ?? '',
+        amount: doc.data['amount'].toDouble() ?? 0,
+        isDeemedPositive: doc.data['isdeemedpositive'].toString() ?? '',
+        isPartyLedger: doc.data['ispartyledger'].toString() ?? '',
+        ledgerGuid: doc.data['ledger_guid'].toString() ?? '',
+        ledgerMasterId: doc.data['ledgername'].toString() ?? '',
+        ledgerRefMasterId: doc.data['ledgerrefname'].toString() ?? '',
+        primaryVoucherType:
+            doc.data['primary_voucher_type_name'].toString() ?? '',
+        primaryGroup: doc.data['primarygroup'].toString() ?? '',
+        ledgerName: doc.data['restat_ledger_name'].toString() ?? '',
+        ledgerRefName: doc.data['restat_ledger_ref_name'].toString() ?? '',
+        partyName: doc.data['restat_party_ledger_name'].toString() ?? '',
+        date: doc.data['voucher_date'].toDate() ?? null,
+        voucherMasterID: doc.data['voucher_master_id'].toString() ?? '',
       );
     }).toList();
   }
